% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_default_aesthetics.R
\docType{data}
\name{all_geoms}
\alias{all_geoms}
\title{A vector of geoms}
\format{
A vector
}
\usage{
all_geoms
}
\description{
A vector of geoms exported from `{ggplot2}` and extension packages. This
vector is iterated on to set the default aesthetics for each geom, allowing
the grattan colours to be plotted automatically.
}
\keyword{datasets}
