% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_helpers.R
\name{ch_add_stacked_labels}
\alias{ch_add_stacked_labels}
\title{Add stacked labels to the right of a chart}
\usage{
ch_add_stacked_labels(
  .plot,
  pct_extend = 1.2,
  nudge_x = NULL,
  segment.colour = NA,
  min.segment.length = 0.1,
  max.overlaps = 20,
  segment.alpha = 0.5,
  segment.size = 0.1,
  hjust = "right",
  direction = "y",
  size = 3,
  ...
)
}
\arguments{
\item{.plot}{(a \code{ggplot2} object) with the \code{label} aesthetic
specified.}

\item{pct_extend}{(numeric, default = \code{1.2}) how much to expand the
x-axis. The existing x-axis range will be multiplied by this value, so the
default expands by 20\%.}

\item{nudge_x}{(numeric; default = NULL) how much to nudge by}

\item{segment.colour}{(character; default = NA) colour of the line segment,
NA means no line segment}

\item{min.segment.length}{Skip drawing segments shorter than this, as unit or
number. Defaults to 0.5. (Default unit is lines, but other units can be
specified by passing \code{unit(x, "units")}).}

\item{max.overlaps}{Exclude text labels that overlap too many things.
Defaults to 10.}

\item{segment.alpha}{(number; default = 0.5) the transparency of the line
segment}

\item{segment.size}{(number; default = 0.1) the size of the line segment}

\item{hjust}{(default = "right") the horizontal justification of the label}

\item{direction}{"both", "x", or "y" -- direction in which to adjust position of labels}

\item{size}{(number, default = 3) the size of the text in the label}

\item{...}{other arguments passed to \code{geom_text_repel}}
}
\value{
a \code{ggplot2} object
}
\description{
This function adds \code{\link[ggrepel]{geom_text_repel}} to a chart. It
requires you to specify the label aesthetic in the original ggplot. It also
extends the x-axis by a certain amount (20\% by default) so the labels have
space. Most of the arguments get passed directly to
\code{\link[ggrepel]{geom_label_repel}}. This expansion will combine with any
other expansion present in the plot, such as arguments based to
\code{scale_x_continuous}. It needs to be used with a \code{\%\>\%} and not a
\code{\+}.
}
\details{
If you get a warning about Aesthetics not being the right length for size,
that's a cryptic message telling you to set the size argument. Setting it to
3 usually looks alright.
}
\examples{
\dontrun{

# not the best examples because they already have stacked labels
# but you'll see what it does
chart_citymapper() \%>\% ch_add_stacked_labels()

chart_jhu_cases() \%>\% ch_add_stacked_labels()

}
}
