% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_helpers.R
\name{colour_text}
\alias{colour_text}
\title{Create the css needed to colour text using ggtext}
\usage{
colour_text(colour, text, is_note = FALSE, bold_labs = TRUE)
}
\arguments{
\item{colour}{(a character) the colour you want to colour text}

\item{text}{(a character) what the coloured text should say}

\item{is_note}{(logical; default = FALSE) is it a note?}

\item{bold_labs}{(logical; default = TRUE) should the labels be bolded?}
}
\value{
(a glue)
}
\description{
This is a helper function to wrap the css to make text certain colours. This
will be useful when making grattan charts. This should be used within `ggtext::geom_richtext()`
or if being used in a sub-title or title it should be used in the `labs()` call together with
`theme(plot.subtitle = element_markdown())`
}
\examples{
\dontrun{library(ggtext)

ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point(size = 3) +
    labs(
        title = glue::glue("**Fisher's *Iris* dataset**
            <span style='font-size:11pt'>Sepal width vs. sepal length for
                                   {colour_text('#0072B2','setosa')}
                                   {colour_text('#D55E00','versicolor')}
                                   {colour_text('#009E73','virginica')}
                                       </span>"),
                                         theme(plot.title = element_markdown(lineheight = 1.1) )
}
}
