% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_helpers.R
\name{deendash}
\alias{deendash}
\title{Remove en-dashes from a data frame}
\usage{
deendash(df)
}
\value{
a data frame
}
\description{
This function is a wrapper around stringr::str_replace_all that removes 
en-dashes from a data frame.
}
