% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_default_aesthetics.R
\name{.set_geom_aesthetics}
\alias{.set_geom_aesthetics}
\title{Set default geom aesthetics}
\usage{
.set_geom_aesthetics(geom, new)
}
\arguments{
\item{geom}{Name of geom/stat to modify (like "point" or "bin"), or a
Geom/Stat object (like GeomPoint or StatBin).}

\item{new}{One of the following:
\itemize{
\item A named list of aesthetics to serve as new defaults.
\item \code{NULL} to reset the defaults.
}}
}
\description{
All arguments are passed to \code{ggplot2::update_geom_defaults}, but this
function adds a \code{NULL} check.
}
