% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_labs.R
\name{extract_labs}
\alias{extract_labs}
\title{Take a ggplot2 object and return a list containing its title, subtitle, and
caption. Recognises Patchwork plots.}
\usage{
extract_labs(p)
}
\arguments{
\item{p}{ggplot2 object}
}
\value{
list with three elements: title, subtitle, caption
}
\description{
Take a ggplot2 object and return a list containing its title, subtitle, and
caption. Recognises Patchwork plots.
}
