% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overleaf_utils.R
\name{find_overleaf_projects}
\alias{find_overleaf_projects}
\title{Find Overleaf projects}
\usage{
find_overleaf_projects(overleaf_base_path)
}
\arguments{
\item{overleaf_base_path}{Path to the Apps/Overleaf folder}
}
\value{
Character vector of project folder names
}
\description{
Find Overleaf projects
}
\keyword{internal}
