% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_utils.R
\name{get_font_folder_path}
\alias{get_font_folder_path}
\title{Get path to shared font files}
\usage{
get_font_folder_path()
}
\value{
Character string path to font folder, or NULL if not found
}
\description{
Get path to shared font files
}
\keyword{internal}
