% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_utils.R
\name{get_font_status_message}
\alias{get_font_status_message}
\title{Get startup message about font configuration}
\usage{
get_font_status_message()
}
\value{
Character string with font status message
}
\description{
Get startup message about font configuration
}
\keyword{internal}
