% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_animation.R
\name{grattan_anim_save}
\alias{grattan_anim_save}
\title{Save gganimate animation with Grattan visual styling}
\usage{
grattan_anim_save(
  filename,
  animation = gganimate::last_animation(),
  type,
  path = NULL,
  nframes = 100,
  ...
)
}
\arguments{
\item{filename}{File to create on disk; required.}

\item{animation}{The animation objkect to save. Defaults to last rendered
animation using `gganimate::last_animation()`}

\item{type}{Either "blog" (to save using the Grattan Blog template) or
"normal" (to save as a Grattan chart as in a report), or
"fullslide_169" (to save as a 16:9 orientation Grattan-branded image).}

\item{path}{Path to save plot to (combined with filename)}

\item{nframes}{Number of frames to include in the saved animation}

\item{...}{arguments passed on to animate}
}
\description{
`grattan_anim_save()` takes a gganimate animation and formats it in line with
the Grattan Institute style guide, in a manner similar to `grattan_save()`
for static plots, then saves the animation to disk.
}
\examples{

library(ggplot2)
library(gganimate)

# First, create a static plot:

static_plot <- ggplot(mtcars, aes(x = wt, y = mpg)) +
 geom_point() +
 labs(title = "Title goes here",
      subtitle = "Subtitle goes here",
      caption = "Notes: some notes. Source: data source") +
 theme_grattan()  +
 theme(axis.title = element_blank())

# Then animate it (this can be in one step with the code above,
# if you'd prefer):

anim_plot <- static_plot +
 transition_states(cyl)

# Then save it. Note that you must specify the chart type. You can optionally
# specify nframes. Other arguments such as
# fps will be passed to gganimate::animate()

\dontrun{grattan_anim_save("test.gif", anim_plot, type = "blog",
fps = 5, nframes = 20)}


}
