% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_animation.R
\name{grattan_animate}
\alias{grattan_animate}
\title{See \code{?gganimate::animate}.
This function is a replica of \code{gganimate::animate}, except for
the addition of a 'type' argument that allows the creation of
Grattan-styled charts.}
\usage{
grattan_animate(
  plot,
  nframes,
  fps,
  duration,
  detail,
  renderer,
  device,
  ref_frame,
  start_pause,
  end_pause,
  rewind,
  type,
  ...
)
}
\arguments{
\item{plot}{A gganim object}

\item{nframes}{The number of frames to render (default 100)}

\item{fps}{The framerate of the animation in frames/sec (default 10)}

\item{duration}{The length of the animation in seconds (unset by default)}

\item{detail}{The number of additional frames to calculate,
per frame (default 1)}

\item{renderer}{The function used to render the generated frames into
an animation. Gets a vector of paths to images along with the framerate.
(default gifski_renderer())}

\item{device}{The device to use for rendering the single frames. Possible
values are 'png', 'jpeg', 'tiff', 'bmp', 'svg', and 'svglite'
(requires the svglite package). (default 'png')}

\item{ref_frame}{The frame to use for fixing dimensions of the plot,
e.g. the space available for axis text. Defaults to the first frame.
Negative values counts backwards (-1 is the last frame) (default 1)}

\item{start_pause, end_pause}{Number of times to repeat the first and
last frame in the animation (default is 0 for both)}

\item{rewind}{Should the animation roll back in the end (default FALSE)}

\item{type}{Either "blog" (to save using the Grattan Blog template) or
"normal" (to save as a Grattan chart as in a report), or
"fullslide_169" (to save as a 16:9 orientation Grattan-branded image).}

\item{...}{Arguments passed on to the device}
}
\description{
See \code{?gganimate::animate}.
This function is a replica of \code{gganimate::animate}, except for
the addition of a 'type' argument that allows the creation of
Grattan-styled charts.
}
