% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_black}
\alias{grattan_black}
\title{Hex code for the colour: Grattan Black (normal black)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_black
}
\description{
#000000
}
\keyword{datasets}
