% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_blue}
\alias{grattan_blue}
\title{Hex code for the colour: Grattan Blue (alias for Grattan Light Blue)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_blue
}
\description{
#A3C7DF
}
\keyword{datasets}
