% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_blue1}
\alias{grattan_blue1}
\title{Hex code for the colour: Grattan Blue [tint 1] (alias for Grattan Light Blue1)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_blue1
}
\description{
#ACCDE2
}
\keyword{datasets}
