% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_blue2}
\alias{grattan_blue2}
\title{Hex code for the colour: Grattan Blue [tint 2] (alias for Grattan Light Blue2)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_blue2
}
\description{
#B5D2E5
}
\keyword{datasets}
