% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_blue3}
\alias{grattan_blue3}
\title{Hex code for the colour: Grattan Blue [tint 3] (alias for Grattan Light Blue3)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_blue3
}
\description{
#BFD8E9
}
\keyword{datasets}
