% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_blue4}
\alias{grattan_blue4}
\title{Hex code for the colour: Grattan Blue [tint 4] (alias for Grattan Light Blue4)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_blue4
}
\description{
#C8DDEC
}
\keyword{datasets}
