% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_blue5}
\alias{grattan_blue5}
\title{Hex code for the colour: Grattan Blue [tint 5] (alias for Grattan Light Blue5)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_blue5
}
\description{
#D1E3EF
}
\keyword{datasets}
