% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_blue6}
\alias{grattan_blue6}
\title{Hex code for the colour: Grattan Blue [tint 6] (alias for Grattan Light Blue6)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_blue6
}
\description{
#DAE9F2
}
\keyword{datasets}
