% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_blue7}
\alias{grattan_blue7}
\title{Hex code for the colour: Grattan Blue [tint 7] (alias for Grattan Light Blue7)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_blue7
}
\description{
#E3EEF5
}
\keyword{datasets}
