% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_blue8}
\alias{grattan_blue8}
\title{Hex code for the colour: Grattan Blue [tint 8] (alias for Grattan Light Blue8)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_blue8
}
\description{
#EDF4F9
}
\keyword{datasets}
