% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_scales.R
\name{grattan_colour_manual}
\alias{grattan_colour_manual}
\title{Convenient functions to set Grattan-appropriate palettes}
\usage{
grattan_colour_manual(n = 0, reverse = FALSE, ...)
}
\arguments{
\item{n}{Numeric. The number of levels in your colour scale.
Minimum value is 1, maximum is 7. Passed to \code{grattan_pal};
see ?grattan_pal for more information.}

\item{reverse}{Logical. FALSE by default. Setting to TRUE reverses the
standard colour order.}

\item{...}{arguments passed to ggplot2 scales}
}
\description{
Convenient functions to set Grattan-appropriate palettes
}
