% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkblue1}
\alias{grattan_darkblue1}
\title{Hex code for the colour: Grattan Dark Blue [tint 1]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkblue1
}
\description{
#538FD4
}
\keyword{datasets}
