% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkblue2}
\alias{grattan_darkblue2}
\title{Hex code for the colour: Grattan Dark Blue [tint 2]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkblue2
}
\description{
#699DD9
}
\keyword{datasets}
