% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkblue4}
\alias{grattan_darkblue4}
\title{Hex code for the colour: Grattan Dark Blue [tint 4]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkblue4
}
\description{
#94B9E4
}
\keyword{datasets}
