% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkblue5}
\alias{grattan_darkblue5}
\title{Hex code for the colour: Grattan Dark Blue [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkblue5
}
\description{
#A9C7E9
}
\keyword{datasets}
