% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkblue7}
\alias{grattan_darkblue7}
\title{Hex code for the colour: Grattan Dark Blue [tint 7]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkblue7
}
\description{
#D4E3F4
}
\keyword{datasets}
