% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkblue8}
\alias{grattan_darkblue8}
\title{Hex code for the colour: Grattan Dark Blue [tint 8]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkblue8
}
\description{
#EAF1FA
}
\keyword{datasets}
