% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkgrey}
\alias{grattan_darkgrey}
\title{Hex code for the colour: Grattan Dark Grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkgrey
}
\description{
#575757
}
\keyword{datasets}
