% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grattan_colours.R
\docType{data}
\name{grattan_darkgrey1}
\alias{grattan_darkgrey1}
\title{'}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkgrey1
}
\description{
'
}
\keyword{datasets}
