% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkgrey2}
\alias{grattan_darkgrey2}
\title{Hex code for the colour: Grattan Dark Grey [tint 2]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkgrey2
}
\description{
#797979
}
\keyword{datasets}
