% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkgrey3}
\alias{grattan_darkgrey3}
\title{Hex code for the colour: Grattan Dark Grey [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkgrey3
}
\description{
#898989
}
\keyword{datasets}
