% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkgrey4}
\alias{grattan_darkgrey4}
\title{Hex code for the colour: Grattan Dark Grey [tint 4]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkgrey4
}
\description{
#9A9A9A
}
\keyword{datasets}
