% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkgrey5}
\alias{grattan_darkgrey5}
\title{Hex code for the colour: Grattan Dark Grey [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkgrey5
}
\description{
#ABABAB
}
\keyword{datasets}
