% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkgrey6}
\alias{grattan_darkgrey6}
\title{Hex code for the colour: Grattan Dark Grey [tint 6]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkgrey6
}
\description{
#BCBCBC
}
\keyword{datasets}
