% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkorange}
\alias{grattan_darkorange}
\title{Hex code for the colour: Grattan Dark Orange (carrot soup)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkorange
}
\description{
#D4582A
}
\keyword{datasets}
