% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkorange2}
\alias{grattan_darkorange2}
\title{Hex code for the colour: Grattan Dark Orange [tint 2]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkorange2
}
\description{
#DD7955
}
\keyword{datasets}
