% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkorange3}
\alias{grattan_darkorange3}
\title{Hex code for the colour: Grattan Dark Orange [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkorange3
}
\description{
#E18A6A
}
\keyword{datasets}
