% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkorange5}
\alias{grattan_darkorange5}
\title{Hex code for the colour: Grattan Dark Orange [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkorange5
}
\description{
#EAAC95
}
\keyword{datasets}
