% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkorange7}
\alias{grattan_darkorange7}
\title{Hex code for the colour: Grattan Dark Orange [tint 7]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkorange7
}
\description{
#F2CDBF
}
\keyword{datasets}
