% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkorange8}
\alias{grattan_darkorange8}
\title{Hex code for the colour: Grattan Dark Orange [tint 8]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkorange8
}
\description{
#F6DED4
}
\keyword{datasets}
