% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkorange_f}
\alias{grattan_darkorange_f}
\title{Hex code for the faded colour: Grattan dark orange (faded)}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_darkorange_f
}
\description{
#DE815F
}
\keyword{datasets}
