% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkred2}
\alias{grattan_darkred2}
\title{Hex code for the colour: Grattan Dark Red [tint 2]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkred2
}
\description{
#814143
}
\keyword{datasets}
