% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkred3}
\alias{grattan_darkred3}
\title{Hex code for the colour: Grattan Dark Red [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkred3
}
\description{
#91595B
}
\keyword{datasets}
