% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkred4}
\alias{grattan_darkred4}
\title{Hex code for the colour: Grattan Dark Red [tint 4]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkred4
}
\description{
#A17172
}
\keyword{datasets}
