% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkred5}
\alias{grattan_darkred5}
\title{Hex code for the colour: Grattan Dark Red [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkred5
}
\description{
#B1898A
}
\keyword{datasets}
