% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkred6}
\alias{grattan_darkred6}
\title{Hex code for the colour: Grattan Dark Red [tint 6]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkred6
}
\description{
#C0A0A1
}
\keyword{datasets}
