% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkred7}
\alias{grattan_darkred7}
\title{Hex code for the colour: Grattan Dark Red [tint 7]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkred7
}
\description{
#D0B8B9
}
\keyword{datasets}
