% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkred8}
\alias{grattan_darkred8}
\title{Hex code for the colour: Grattan Dark Red [tint 8]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkred8
}
\description{
#E0D0D0
}
\keyword{datasets}
