% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkred_f}
\alias{grattan_darkred_f}
\title{Hex code for the faded colour: Grattan dark red (faded)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkred_f
}
\description{
#894D4E
}
\keyword{datasets}
