% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkyellow}
\alias{grattan_darkyellow}
\title{Hex code for the colour: Grattan Dark Yellow (butternut pumpkin soup)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkyellow
}
\description{
#FFC35A
}
\keyword{datasets}
