% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkyellow1}
\alias{grattan_darkyellow1}
\title{Hex code for the colour: Grattan Dark Yellow [tint 1]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkyellow1
}
\description{
#FFC96B
}
\keyword{datasets}
