% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkyellow3}
\alias{grattan_darkyellow3}
\title{Hex code for the colour: Grattan Dark Yellow [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkyellow3
}
\description{
#FFD58C
}
\keyword{datasets}
