% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkyellow4}
\alias{grattan_darkyellow4}
\title{Hex code for the colour: Grattan Dark Yellow [tint 4]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkyellow4
}
\description{
#FFDB9C
}
\keyword{datasets}
