% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkyellow5}
\alias{grattan_darkyellow5}
\title{Hex code for the colour: Grattan Dark Yellow [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkyellow5
}
\description{
#FFE1AD
}
\keyword{datasets}
