% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkyellow6}
\alias{grattan_darkyellow6}
\title{Hex code for the colour: Grattan Dark Yellow [tint 6]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkyellow6
}
\description{
#FFE7BD
}
\keyword{datasets}
