% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_grey1}
\alias{grattan_grey1}
\title{Hex code for the colour: Grattan grey 1 (truffle soup)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_grey1
}
\description{
#D9D9D9
}
\keyword{datasets}
