% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_grey2}
\alias{grattan_grey2}
\title{Hex code for the colour: Grattan grey 2 (cream of mushroom soup)}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_grey2
}
\description{
#AEAEAE
}
\keyword{datasets}
