% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_grey3}
\alias{grattan_grey3}
\title{Hex code for the colour: Grattan grey 3 (morel mushroom soup)}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_grey3
}
\description{
#828282
}
\keyword{datasets}
