% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_grey4}
\alias{grattan_grey4}
\title{Hex code for the colour: Grattan grey 4 (shiitake mushroom soup)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_grey4
}
\description{
#575757
}
\keyword{datasets}
