% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_grey5}
\alias{grattan_grey5}
\title{Hex code for the colour: Grattan grey 5 (blackbean soup)}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_grey5
}
\description{
#2B2B2B
}
\keyword{datasets}
