% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_grey_alpha}
\alias{grattan_grey_alpha}
\title{Hex code for the colour: Grattan 'forecast shading' grey}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_grey_alpha
}
\description{
#E1E3E5
}
\keyword{datasets}
