% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grattan_colours.R
\docType{data}
\name{grattan_grey_title}
\alias{grattan_grey_title}
\title{grattan_grey_title}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_grey_title
}
\description{
grattan_grey_title
}
\keyword{datasets}
