% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_gridlinegrey}
\alias{grattan_gridlinegrey}
\title{Hex code for the grey used for gridlines in Grattan charts}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_gridlinegrey
}
\description{
#C3C7CB
}
\keyword{datasets}
