% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightblue}
\alias{grattan_lightblue}
\title{Hex code for the colour: Grattan Light Blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightblue
}
\description{
#A3C7DF
}
\keyword{datasets}
