% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightblue1}
\alias{grattan_lightblue1}
\title{Hex code for the colour: Grattan Light Blue [tint 1]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightblue1
}
\description{
#ACCDE2
}
\keyword{datasets}
