% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightblue2}
\alias{grattan_lightblue2}
\title{Hex code for the colour: Grattan Light Blue [tint 2]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightblue2
}
\description{
#B5D2E5
}
\keyword{datasets}
