% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightblue3}
\alias{grattan_lightblue3}
\title{Hex code for the colour: Grattan Light Blue [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightblue3
}
\description{
#BFD8E9
}
\keyword{datasets}
