% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightblue4}
\alias{grattan_lightblue4}
\title{Hex code for the colour: Grattan Light Blue [tint 4]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightblue4
}
\description{
#C8DDEC
}
\keyword{datasets}
