% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightblue6}
\alias{grattan_lightblue6}
\title{Hex code for the colour: Grattan Light Blue [tint 6]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightblue6
}
\description{
#DAE9F2
}
\keyword{datasets}
