% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightblue7}
\alias{grattan_lightblue7}
\title{Hex code for the colour: Grattan Light Blue [tint 7]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightblue7
}
\description{
#E3EEF5
}
\keyword{datasets}
