% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightblue8}
\alias{grattan_lightblue8}
\title{Hex code for the colour: Grattan Light Blue [tint 8]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightblue8
}
\description{
#EDF4F9
}
\keyword{datasets}
