% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightgrey1}
\alias{grattan_lightgrey1}
\title{Hex code for the colour: Grattan Light Grey [tint 1]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightgrey1
}
\description{
#8F8F8F
}
\keyword{datasets}
