% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightgrey2}
\alias{grattan_lightgrey2}
\title{Hex code for the colour: Grattan Light Grey [tint 2]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightgrey2
}
\description{
#9B9B9B
}
\keyword{datasets}
