% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightgrey3}
\alias{grattan_lightgrey3}
\title{Hex code for the colour: Grattan Light Grey [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightgrey3
}
\description{
#A8A8A8
}
\keyword{datasets}
