% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightgrey4}
\alias{grattan_lightgrey4}
\title{Hex code for the colour: Grattan Light Grey [tint 4]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightgrey4
}
\description{
#B4B4B4
}
\keyword{datasets}
