% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightgrey5}
\alias{grattan_lightgrey5}
\title{Hex code for the colour: Grattan Light Grey [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightgrey5
}
\description{
#C1C1C1
}
\keyword{datasets}
