% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightgrey6}
\alias{grattan_lightgrey6}
\title{Hex code for the colour: Grattan Light Grey [tint 6]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightgrey6
}
\description{
#CDCDCD
}
\keyword{datasets}
