% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightgrey8}
\alias{grattan_lightgrey8}
\title{Hex code for the colour: Grattan Light Grey [tint 8]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightgrey8
}
\description{
#E6E6E6
}
\keyword{datasets}
