% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange}
\alias{grattan_lightorange}
\title{Hex code for the colour: Grattan Light Orange (sweet potato soup)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightorange
}
\description{
#F68B33
}
\keyword{datasets}
