% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange1}
\alias{grattan_lightorange1}
\title{Hex code for the colour: Grattan Light Orange [tint 1]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightorange1
}
\description{
#F79747
}
\keyword{datasets}
