% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange2}
\alias{grattan_lightorange2}
\title{Hex code for the colour: Grattan Light Orange [tint 2]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightorange2
}
\description{
#F8A25C
}
\keyword{datasets}
