% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange3}
\alias{grattan_lightorange3}
\title{Hex code for the colour: Grattan Light Orange [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightorange3
}
\description{
#F9AE70
}
\keyword{datasets}
