% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange4}
\alias{grattan_lightorange4}
\title{Hex code for the colour: Grattan Light Orange [tint 4]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightorange4
}
\description{
#FAB985
}
\keyword{datasets}
