% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange5}
\alias{grattan_lightorange5}
\title{Hex code for the colour: Grattan Light Orange [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightorange5
}
\description{
#FBC599
}
\keyword{datasets}
