% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange6}
\alias{grattan_lightorange6}
\title{Hex code for the colour: Grattan Light Orange [tint 6]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightorange6
}
\description{
#FBD1AD
}
\keyword{datasets}
