% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange7}
\alias{grattan_lightorange7}
\title{Hex code for the colour: Grattan Light Orange [tint 7]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightorange7
}
\description{
#FCDCC2
}
\keyword{datasets}
