% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange8}
\alias{grattan_lightorange8}
\title{Hex code for the colour: Grattan Light Orange [tint 8]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightorange8
}
\description{
#FDE8D6
}
\keyword{datasets}
