% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightorange_f}
\alias{grattan_lightorange_f}
\title{Hex code for the faded colour: Grattan light orange (faded)}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_lightorange_f
}
\description{
#F8A866
}
\keyword{datasets}
