% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightred}
\alias{grattan_lightred}
\title{Hex code for the colour: Grattan Light Red (tomato soup)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightred
}
\description{
#A02226
}
\keyword{datasets}
