% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightred1}
\alias{grattan_lightred1}
\title{Hex code for the colour: Grattan Light Red [tint 1]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightred1
}
\description{
#AA383C
}
\keyword{datasets}
