% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightred2}
\alias{grattan_lightred2}
\title{Hex code for the colour: Grattan Light Red [tint 2]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightred2
}
\description{
#B34E51
}
\keyword{datasets}
