% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightred3}
\alias{grattan_lightred3}
\title{Hex code for the colour: Grattan Light Red [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightred3
}
\description{
#BD6467
}
\keyword{datasets}
