% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightred4}
\alias{grattan_lightred4}
\title{Hex code for the colour: Grattan Light Red [tint 4]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightred4
}
\description{
#C67A7D
}
\keyword{datasets}
