% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightred6}
\alias{grattan_lightred6}
\title{Hex code for the colour: Grattan Light Red [tint 6]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightred6
}
\description{
#D9A7A8
}
\keyword{datasets}
