% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightred8}
\alias{grattan_lightred8}
\title{Hex code for the colour: Grattan Light Red [tint 8]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightred8
}
\description{
#ECD3D4
}
\keyword{datasets}
