% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow}
\alias{grattan_lightyellow}
\title{Hex code for the colour: Grattan light yello (corn chowder)}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_lightyellow
}
\description{
#FFE07F
}
\keyword{datasets}
