% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow1}
\alias{grattan_lightyellow1}
\title{Hex code for the colour: Grattan Light Yellow [tint 1]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightyellow1
}
\description{
#FFE38C
}
\keyword{datasets}
