% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow4}
\alias{grattan_lightyellow4}
\title{Hex code for the colour: Grattan Light Yellow [tint 4]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightyellow4
}
\description{
#FFECB2
}
\keyword{datasets}
