% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow5}
\alias{grattan_lightyellow5}
\title{Hex code for the colour: Grattan Light Yellow [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightyellow5
}
\description{
#FFF0BF
}
\keyword{datasets}
