% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grattan_colours.R
\docType{data}
\name{grattan_lightyellow6}
\alias{grattan_lightyellow6}
\title{grattan_lightyellow6}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightyellow6
}
\description{
grattan_lightyellow6
}
\keyword{datasets}
