% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow7}
\alias{grattan_lightyellow7}
\title{Hex code for the colour: Grattan Light Yellow [tint 7]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightyellow7
}
\description{
#FFF6D9
}
\keyword{datasets}
