% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow8}
\alias{grattan_lightyellow8}
\title{Hex code for the colour: Grattan Light Yellow [tint 8]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightyellow8
}
\description{
#FFF9E5
}
\keyword{datasets}
