% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow_f}
\alias{grattan_lightyellow_f}
\title{Hex code for the faded colour: Grattan light yellow (faded)}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_lightyellow_f
}
\description{
#FFE79F
}
\keyword{datasets}
