% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_orange}
\alias{grattan_orange}
\title{Hex code for the colour: Grattan light orange (sweet potato soup)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_orange
}
\description{
#F68B33
}
\keyword{datasets}
