% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_orange1}
\alias{grattan_orange1}
\title{Hex code for the colour: Grattan Orange [tint 1] (alias for Grattan Light Orange1)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_orange1
}
\description{
#F79747
}
\keyword{datasets}
