% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_orange3}
\alias{grattan_orange3}
\title{Hex code for the colour: Grattan Orange [tint 3] (alias for Grattan Light Orange3)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_orange3
}
\description{
#F9AE70
}
\keyword{datasets}
