% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_orange4}
\alias{grattan_orange4}
\title{Hex code for the colour: Grattan Orange [tint 4] (alias for Grattan Light Orange4)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_orange4
}
\description{
#FAB985
}
\keyword{datasets}
