% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_orange5}
\alias{grattan_orange5}
\title{Hex code for the colour: Grattan Orange [tint 5] (alias for Grattan Light Orange5)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_orange5
}
\description{
#FBC599
}
\keyword{datasets}
