% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_orange6}
\alias{grattan_orange6}
\title{Hex code for the colour: Grattan Orange [tint 6] (alias for Grattan Light Orange6)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_orange6
}
\description{
#FBD1AD
}
\keyword{datasets}
