% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_orange7}
\alias{grattan_orange7}
\title{Hex code for the colour: Grattan Orange [tint 7] (alias for Grattan Light Orange7)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_orange7
}
\description{
#FCDCC2
}
\keyword{datasets}
