% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_orange8}
\alias{grattan_orange8}
\title{Hex code for the colour: Grattan Orange [tint 8] (alias for Grattan Light Orange8)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_orange8
}
\description{
#FDE8D6
}
\keyword{datasets}
