% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_orange_alpha}
\alias{grattan_orange_alpha}
\title{Hex code for the colour used for Grattan box backgrounds (orange alpha)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_orange_alpha
}
\description{
#FEF0DE
}
\keyword{datasets}
