% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_pal.R
\name{grattan_pal}
\alias{grattan_pal}
\title{Create a Grattan-appropriate palette for your chart.}
\usage{
grattan_pal(n, reverse = FALSE)
}
\arguments{
\item{n}{The number of levels in your colour scale. Acceptable values are between 1 and 6.}

\item{reverse}{FALSE by default. Non-reverse order runs from light to dark.}
}
\description{
Create a Grattan-appropriate palette for your chart.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    theme_grattan() +
    scale_colour_manual(grattan_pal(n = 3))

}
