% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_pal.R
\name{grattan_palettes}
\alias{grattan_palettes}
\title{Predefined grattan colours combined into palettes}
\usage{
grattan_palettes
}
\description{
This is a list of grattan colours combined into palettes. The palettes are used
for different plots and maps.
}
