% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_red}
\alias{grattan_red}
\title{Hex code for the colour: Grattan Red (alias for Grattan Light Red)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_red
}
\description{
#A02226
}
\keyword{datasets}
