% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_red1}
\alias{grattan_red1}
\title{Hex code for the colour: Grattan Red [tint 1] (alias for Grattan Light Red1)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_red1
}
\description{
#AA383C
}
\keyword{datasets}
