% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_red3}
\alias{grattan_red3}
\title{Hex code for the colour: Grattan Red [tint 3] (alias for Grattan Light Red3)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_red3
}
\description{
#BD6467
}
\keyword{datasets}
