% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_red4}
\alias{grattan_red4}
\title{Hex code for the colour: Grattan Red [tint 4] (alias for Grattan Light Red4)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_red4
}
\description{
#C67A7D
}
\keyword{datasets}
