% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_red5}
\alias{grattan_red5}
\title{Hex code for the colour: Grattan Red [tint 5] (alias for Grattan Light Red5)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_red5
}
\description{
#D09193
}
\keyword{datasets}
