% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_red6}
\alias{grattan_red6}
\title{Hex code for the colour: Grattan Red [tint 6] (alias for Grattan Light Red6)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_red6
}
\description{
#D9A7A8
}
\keyword{datasets}
