% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_red8}
\alias{grattan_red8}
\title{Hex code for the colour: Grattan Red [tint 8] (alias for Grattan Light Red8)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_red8
}
\description{
#ECD3D4
}
\keyword{datasets}
