% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_red_f}
\alias{grattan_red_f}
\title{Hex code for the faded colour: Grattan red (faded)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_red_f
}
\description{
#B7595C
}
\keyword{datasets}
