% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_save_pptx.R
\name{grattan_save_pptx_onetype}
\alias{grattan_save_pptx_onetype}
\title{Internal function to create a Grattan-branded PPTX slide using RMd
Called by grattan_save_pptx(). This function saves slide(s) in a single
format.}
\usage{
grattan_save_pptx_onetype(p, filename, type)
}
\arguments{
\item{p}{ggplot2 plot or list of plots}

\item{filename}{filename (incl. path where necessary) to save file}

\item{type}{type of slide; see \code{?grattan_save_pptx()} for types}
}
\description{
Internal function to create a Grattan-branded PPTX slide using RMd
Called by grattan_save_pptx(). This function saves slide(s) in a single
format.
}
