% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_yellow1}
\alias{grattan_yellow1}
\title{Hex code for the colour: Grattan Yellow [tint 1] (alias for Grattan Dark Yellow1)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_yellow1
}
\description{
#FFC96B
}
\keyword{datasets}
