% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_yellow2}
\alias{grattan_yellow2}
\title{Hex code for the colour: Grattan Yellow [tint 2] (alias for Grattan Dark Yellow2)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_yellow2
}
\description{
#FFCF7B
}
\keyword{datasets}
