% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_yellow3}
\alias{grattan_yellow3}
\title{Hex code for the colour: Grattan Yellow [tint 3] (alias for Grattan Dark Yellow3)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_yellow3
}
\description{
#FFD58C
}
\keyword{datasets}
