% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_yellow4}
\alias{grattan_yellow4}
\title{Hex code for the colour: Grattan Yellow [tint 4] (alias for Grattan Dark Yellow4)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_yellow4
}
\description{
#FFDB9C
}
\keyword{datasets}
