% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_yellow5}
\alias{grattan_yellow5}
\title{Hex code for the colour: Grattan Yellow [tint 5] (alias for Grattan Dark Yellow5)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_yellow5
}
\description{
#FFE1AD
}
\keyword{datasets}
