% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_yellow6}
\alias{grattan_yellow6}
\title{Hex code for the colour: Grattan Yellow [tint 6] (alias for Grattan Dark Yellow6)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_yellow6
}
\description{
#FFE7BD
}
\keyword{datasets}
