% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_yellow7}
\alias{grattan_yellow7}
\title{Hex code for the colour: Grattan Yellow [tint 7] (alias for Grattan Dark Yellow7)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_yellow7
}
\description{
#FFEDCE
}
\keyword{datasets}
