% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_yellow8}
\alias{grattan_yellow8}
\title{Hex code for the colour: Grattan Yellow [tint 8] (alias for Grattan Dark Yellow8)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_yellow8
}
\description{
#FFF3DE
}
\keyword{datasets}
