% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_yellow_f}
\alias{grattan_yellow_f}
\title{Hex code for the faded colour: Grattan yellow (faded)}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_yellow_f
}
\description{
#FFD283
}
\keyword{datasets}
