% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattantheme-package.R
\docType{package}
\name{grattantheme-package}
\alias{grattantheme}
\alias{grattantheme-package}
\title{grattantheme: Create Graphs in the Grattan Institute Style}
\description{
Enables the user to create ggplot2 charts that conform to the Grattan Institute style guide.
}
\author{
\strong{Maintainer}: Will Mackey \email{william.mackey@grattaninstitute.edu.au}

Authors:
\itemize{
  \item Matt Cowgill \email{matthew.cowgill@grattaninstitute.edu.au}
}

}
\keyword{internal}
