% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridextra_fullslide.R
\name{gridextra_fullslide}
\alias{gridextra_fullslide}
\title{Legacy function, used for animation, because you can't animate patchwork plots}
\usage{
gridextra_fullslide(
  object,
  type = "fullslide",
  height = NULL,
  print_object = FALSE
)
}
\description{
Legacy function, used for animation, because you can't animate patchwork plots
}
\keyword{internal}
