% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_pal.R
\name{make_grattan_pal}
\alias{make_grattan_pal}
\title{Interpolate a grattan colour palette}
\usage{
make_grattan_pal(palette = "sequential", reverse = FALSE, ...)
}
\arguments{
\item{palette}{(character; default = \code{"graph"}) given name of a grattan
palette: \code{\link{grattan_palettes}}}

\item{reverse}{(boolean; default = \code{FALSE}) indicating if palette should
be reverse}

\item{...}{Additional arguments to pass to \code{colorRampPalette} see
details here \code{\link[grDevices]{colorRamp}}}
}
\description{
This function takes a grattan colour palette and generates more colours from it,
so that there are enough to make your chart.
}
\details{
The interpolation method is set to "spline" (the default is "linear") in an
attempt to reduce the number of vomit colours that get produced when
generating many colours.

It returns a function that takes a single value and makes that many colours.
}
\examples{

ggplot2::ggplot(mtcars, ggplot2::aes(x = mpg, y = hp, colour = as.character(wt))) +
  ggplot2::geom_point() +
  ggplot2::scale_colour_manual(values = make_grattan_pal()(29))

}
\seealso{
\code{\link{grattan_palettes}}
}
