% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_pal.R
\name{make_grattan_pal_discrete}
\alias{make_grattan_pal_discrete}
\title{Create a grattan colour palette}
\usage{
make_grattan_pal_discrete(n)
}
\arguments{
\item{n}{how many colours to return}
}
\description{
This function takes a the grattan graph colour palette and returns a vector of colours equal to n.
It is used in \code{\link{scale_colour_grattan}} and \code{\link{scale_fill_grattan}} to make the discrete
colour scale as the order of colours is specific in the grattan branding guides and so using an interpolated scale
does not work.
}
\seealso{
\code{\link{grattan_palettes}}
}
