% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_slide.R
\name{make_slide}
\alias{make_slide}
\title{Create a Powerpoint slide that contains a Grattan graph and editable title and subtitle}
\usage{
make_slide(graph = last_plot(), filename = NULL, path = ".",
  type = "16:9")
}
\arguments{
\item{graph}{The graph you want to include on your slide. Defaults to the last plot (ggplot2::last_plot())}

\item{filename}{The filename for your Powerpoint slide.}

\item{path}{Path to the directory you wish to save your slide in. Defaults to your working directory.}

\item{type}{The type of Powerpoint slide to create; either "16:9" (the default, for a 16:9 orientation slide), or "4:3".}
}
\description{
Create a Powerpoint slide that contains a Grattan graph and editable title and subtitle
}
