% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_pal.R
\name{palette_order}
\alias{palette_order}
\title{Set the palette order as needed in the}
\usage{
palette_order
}
\description{
This is a list of grattan colours combined into palettes. The palettes are used
for different plots and maps.
}
