% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_save_pptx.R
\name{pandoc_test}
\alias{pandoc_test}
\title{Check that pandoc is available and is an appropriate version}
\usage{
pandoc_test()
}
\description{
Check that pandoc is available and is an appropriate version
}
\keyword{internal}
