% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_helpers.R
\name{paulify_classifications}
\alias{paulify_classifications}
\title{Make the ANZSIC and ANZSCO classifications Paul-compliant}
\usage{
paulify_classifications(classification, remove_services = FALSE)
}
\arguments{
\item{classification}{(character) the industry classification to be paulified}

\item{remove_services}{(logical) whether to remove the word services in the classification}
}
\value{
a character
}
\description{
This inserts oxford commas and and sorts out the capitalisation so that ANZSCO and
ANZSIC classifications are compliance with Paul's standards.
}
\examples{

\dontrun{
    strayr::anzsic2006  \%>\%
       mutate(industry = paulify_classifications(anzsic_division))
}

}
