% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_default_aesthetics.R
\name{plot_opts_grattan}
\alias{plot_opts_grattan}
\title{A list of plot options reflecting the Grattan brand}
\usage{
plot_opts_grattan
}
\description{
A list of plot options reflecting the Grattan brand
}
