% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_default_aesthetics.R
\name{plot_opts_vanilla}
\alias{plot_opts_vanilla}
\title{The vanilla plotting options}
\usage{
plot_opts_vanilla
}
\description{
These options restore the vanilla ggplot2 colours options for the R session.
}
