% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\docType{data}
\name{qnd_plot_opts}
\alias{qnd_plot_opts}
\title{The QND plotting options}
\format{
An object of class \code{list} of length 4.
}
\usage{
qnd_plot_opts
}
\description{
These options set the default ggplot2 colours options for the R session.
}
\keyword{datasets}
