% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_pal.R
\name{register_palette}
\alias{register_palette}
\title{Register the option for which palette to use.}
\usage{
register_palette()
}
\description{
Grattan now has an old palette and a new palette. Depending which is needed users can
set options("grattan_palette" = "old") to have the colours map to the old colours. Otherwise the colours
default to the new colours.
}
