% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pptx_error.R
\name{remove_pptx_error}
\alias{remove_pptx_error}
\title{Remove 'image could not be displayed' box from Powerpoint title slide}
\usage{
remove_pptx_error(filename)
}
\arguments{
\item{filename}{the filename of the .pptx file to be fixed}
}
\description{
When creating .pptx presentations with a title slide using
using Pandoc / RMarkdown, the image on the title slide has an 'image cannot
be displayed' box over the image. This function removes the error box. This
function is not intended to be called directly; it is called at the end of
`make_presentation()` where a title slide is present.

This is a temporary fix based on: https://github.com/jgm/pandoc/issues/4827#issuecomment-463596185
}
