% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_labs.R
\name{replace_labs}
\alias{replace_labs}
\title{Replace a ggplot2's labels (title, subtitle, and/or caption) with a
given string. Works for regular ggplot2 plots and patchwork plots.}
\usage{
replace_labs(p, labs = list(title = NULL, subtitle = NULL, caption = NULL))
}
\arguments{
\item{p}{ggplot2 object}

\item{labs}{named list of labels to replace; must contain elements named
'title', 'subtitle', and 'caption', and no other elements.}
}
\description{
Replace a ggplot2's labels (title, subtitle, and/or caption) with a
given string. Works for regular ggplot2 plots and patchwork plots.
}
