% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_scales.R
\name{scale_colour_grattan}
\alias{scale_colour_grattan}
\alias{scale_fill_grattan}
\title{Scale constructor for grattan colours}
\usage{
scale_colour_grattan(palette = NULL, discrete = FALSE, reverse = FALSE, ...)

scale_fill_grattan(palette = NULL, discrete = FALSE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{grattan_palettes}}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or
not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to \code{\link{discrete_scale}} or
\code{\link{scale_color_gradientn}}, used respectively when discrete is
TRUE or FALSE}
}
\description{
If you get an error about not enough colours, try a different palette (such
as \code{graph}) or set \code{discrete = FALSE}. If you are plotting discrete colours
and need more than 10 colours, then you will have to make your own palette using
\code{\link{make_grattan_pal}}.
}
\section{Functions}{
\itemize{
\item \code{scale_colour_grattan()}: For colour scales

\item \code{scale_fill_grattan()}: For fill scales

}}
