% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_default_aesthetics.R
\name{set_aesthetics}
\alias{set_aesthetics}
\title{Set \code{ggplot2} default aesthetics}
\usage{
set_aesthetics(type)
}
\arguments{
\item{type}{(character) Which aesthetics to use? One of "grattan" or "vanilla".}
}
\description{
Sets global options for \code{ggplot2}. If \code{type = "grattan"}, grattan brand
elements (colours and fonts) will be used by default in ggplot. To restore
the defaults use \code{type = "vanilla"}.
}
\examples{
\dontrun{
set_aesthetics(type = "vanilla")
}
}
