% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\name{set_plot_colours}
\alias{set_plot_colours}
\title{Set \code{ggplot2} default colours}
\usage{
set_plot_colours(type)
}
\arguments{
\item{type}{(character) Which colours to use? One of "qnd" or "default".}
}
\description{
Sets global options for \code{ggplot2}. If \code{type = "qnd"}, QND colours
will be used by default in ggplot. If there are #' more than 32 levels it
will switch to \code{viridis} by default, so then you must use
\code{scale_colour_qnd}. To restore the defaults use \code{type = "default"}.
}
\examples{
\dontrun{
set_plot_colours(type = "default")
}
}
