% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\name{set_qnd_defaults}
\alias{set_qnd_defaults}
\title{Set the default ggplot2 aesthetics to QND's}
\usage{
set_qnd_defaults()
}
\description{
A wrapper for a bunch of other functions to set ggplot2 default aesthetics.
Used on load.
}
