% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch_pdfs.R
\name{stitch_pdfs}
\alias{stitch_pdfs}
\title{Combine multiple PDFs into a single PDF}
\usage{
stitch_pdfs(pdfs, output = NULL)
}
\arguments{
\item{pdfs}{character vector containing filenames (with path, if the files
aren't in the working directory) of the PDFs to be combined}

\item{output}{name of the PDF file (including path where necessary)
to be created, containing the stiched-together PDFs}
}
\description{
Deprecated function, no longer available.
Consider using pdftools::pdf_combine() to combine multiple PDFs.
}
