% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_helpers.R
\name{str_wrap_factor}
\alias{str_wrap_factor}
\title{Wrap strings without dropping factor levels}
\usage{
str_wrap_factor(x, ...)
}
\value{
a factor
}
\description{
This function is a wrapper around stringr::str_wrap that ensures that
factor levels are not dropped.
}
