% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grattan.R
\name{theme_grattan}
\alias{theme_grattan}
\title{Create a ggplot2 theme consistent with the Grattan style guide.}
\arguments{
\item{base_size}{Size for text elements. Defaults to 18.}

\item{base_family}{Font family for text elements. Defaults to "sans", indistinguishable from Arial.}
}
\description{
Create a ggplot2 theme consistent with the Grattan style guide.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    theme_grattan()
p
}
