% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark}
\alias{watermark}
\title{Add a watermark annotation layer for a ggplot2 object}
\usage{
watermark(
  watermark,
  fontsize = 120,
  colour = "grey90",
  alpha = 0.1,
  fontface = "bold",
  angle = 22
)
}
\arguments{
\item{watermark}{String to be added as watermark}

\item{fontsize}{Font size}

\item{colour}{Font colour}

\item{alpha}{Alpha (transparency; lower number = more transparent)}

\item{fontface}{Font face ("bold" by default)}

\item{angle}{Angle of the watermark}
}
\description{
Add a watermark annotation layer for a ggplot2 object
}
\examples{
# First, start with a plot:
p <- ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point() + theme_grattan()

# Then add a watermark:

p + watermark("DRAFT")
}
