% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\docType{data}
\name{all_geoms}
\alias{all_geoms}
\title{All ggplot geoms}
\format{
An object of class \code{list} of length 44.
}
\usage{
all_geoms
}
\description{
A list of ggplot2 geoms for restoring default aesthetics
}
\keyword{datasets}
