% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_helper.R
\name{check_chart_aspect_ratio}
\alias{check_chart_aspect_ratio}
\title{Check that your chart looks alright in the different chart formats}
\usage{
check_chart_aspect_ratio(object = ggplot2::last_plot(), type = "normal")
}
\arguments{
\item{object}{The ggplot2 graph object to be saved. Defaults to
\code{last_plot()}, which will save the last plot that was displayed in
your session.}

\item{type}{Sets height and width to Grattan defaults. The following chart
  types are available:

\itemize{
  \item{"normal"}{ The default. Use for normal Grattan report charts,
   or to paste into a 4:3 Powerpoint slide. Width: 22.2cm, height: 14.5cm.}
  \item{"normal_169"}{ Useful for pasting into a 16:9 format Grattan
  Powerpoint slide. Width: 30cm, height: 14.5cm.}
  \item{"tiny"}{ Fills the width of a column in a Grattan report,
  but is shorter than usual. Width: 22.2cm, height: 11.1cm.}
  \item{"wholecolumn"}{ Takes up a whole column in a Grattan report.
  Width: 22.2cm, height: 22.2cm.}
  \item{"fullpage"}{ Fills a whole page of a Grattan report.
  Width: 44.3cm, height: 22.2cm.}
  \item{"fullslide}{ A 16:9 Grattan
  Powerpoint slide, complete with logo. Use this to drop into standard
  presentations. Width: 33.9cm, height: 19.0cm}
  \item{"blog"}{"Creates a 4:3 image that looks like a Grattan Powerpoint
  slide, but with less border whitespace than `fullslide`."}
}
Set type = "all" to save your chart in all available sizes or use
`grattan_save_all()`.}
}
\description{
This function will open your plot in a new window with the aspect ratio of the
chart type selected
}
\examples{
\dontrun{
ggplot(mtcars, aes(x = wt, y = mpg)) + geom_point()
check_chart_aspect_ratio()}

}
