% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fullslide.R
\name{create_fullslide}
\alias{create_fullslide}
\title{Assemble a chart featuring the Grattan logo and orange line}
\usage{
create_fullslide(plot = last_plot(), type)
}
\arguments{
\item{plot}{A ggplot2 plot}

\item{type}{Optional. If specified, must be one of "fullslide", "fullslide_169", "fullslide_44", or "blog".
This is used to define the size of the white border around the image.}
}
\value{
An object of class "patchwork".
}
\description{
Takes a ggplot2 object and formats it to look like a
Grattan Powerpoint slide. You will rarely need to call this function
directly - use `grattan_save()` to save a ggplot2 object as a 'slide'-like
image.
}
\examples{

library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    labs(title = "My title",
         subtitle = "My subtitle",
         caption = "My caption") +
    theme_grattan()

# Create an image that includes the Grattan logo

p_logo <- create_fullslide(p)


}
