% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_latex_code.R
\name{export_latex_code}
\alias{export_latex_code}
\title{Export the figure environment LaTeX code for a ggplot2 graph object.}
\usage{
export_latex_code(
  p = ggplot2::last_plot(),
  chart_path = "atlas/chart.pdf",
  code_to_clipboard = TRUE
)
}
\arguments{
\item{p}{The ggplot2 graph object to be saved. Defaults to
\code{last_plot()}, which will save the last plot that was displayed in
your session.}

\item{chart_path}{The save path of the chart. The filename will be extracted
and used in the \code{\\includegraphics} call.}

\item{code_to_clipboard}{Logical, \code{TRUE} by default. If \code{TRUE}, the LaTeX code
to create the figure environment will be copied to your clipboard.}
}
\value{
LaTeX code to create a figure environment
}
\description{
Export the figure environment LaTeX code for a ggplot2 graph object.
}
