% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkblue}
\alias{grattan_darkblue}
\title{Hex code for the colour: Grattan Dark Blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkblue
}
\description{
#3E81CE
}
\keyword{datasets}
