% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkblue3}
\alias{grattan_darkblue3}
\title{Hex code for the colour: Grattan Dark Blue [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkblue3
}
\description{
#7EABDF
}
\keyword{datasets}
