% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkgrey1}
\alias{grattan_darkgrey1}
\title{Hex code for the colour: Grattan Dark Grey [tint 1]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkgrey1
}
\description{
#686868
}
\keyword{datasets}
