% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkred}
\alias{grattan_darkred}
\title{Hex code for the colour: Grattan dark red (kidney bean soup)}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_darkred
}
\description{
#621214
}
\keyword{datasets}
