% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_darkyellow8}
\alias{grattan_darkyellow8}
\title{Hex code for the colour: Grattan Dark Yellow [tint 8]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_darkyellow8
}
\description{
#FFF3DE
}
\keyword{datasets}
