% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_grey_title}
\alias{grattan_grey_title}
\title{Hex code for the colour: Grattan title grey}
\format{An object of class \code{character} of length 1.}
\usage{
grattan_grey_title
}
\description{
#6A737B
}
\keyword{datasets}
