% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightblue5}
\alias{grattan_lightblue5}
\title{Hex code for the colour: Grattan Light Blue [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightblue5
}
\description{
#D1E3EF
}
\keyword{datasets}
