% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightgrey7}
\alias{grattan_lightgrey7}
\title{Hex code for the colour: Grattan Light Grey [tint 7]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightgrey7
}
\description{
#DADADA
}
\keyword{datasets}
