% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightred5}
\alias{grattan_lightred5}
\title{Hex code for the colour: Grattan Light Red [tint 5]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightred5
}
\description{
#D09193
}
\keyword{datasets}
