% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightred7}
\alias{grattan_lightred7}
\title{Hex code for the colour: Grattan Light Red [tint 7]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightred7
}
\description{
#E3BDBE
}
\keyword{datasets}
