% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow2}
\alias{grattan_lightyellow2}
\title{Hex code for the colour: Grattan Light Yellow [tint 2]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightyellow2
}
\description{
#FFE699
}
\keyword{datasets}
