% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow3}
\alias{grattan_lightyellow3}
\title{Hex code for the colour: Grattan Light Yellow [tint 3]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightyellow3
}
\description{
#FFE9A5
}
\keyword{datasets}
