% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{grattan_lightyellow6}
\alias{grattan_lightyellow6}
\title{Hex code for the colour: Grattan Light Yellow [tint 6]}
\format{
An object of class \code{character} of length 1.
}
\usage{
grattan_lightyellow6
}
\description{
#FFF3CC
}
\keyword{datasets}
