% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_pal.R
\name{grattan_pal}
\alias{grattan_pal}
\title{Create a Grattan-appropriate palette for your chart.}
\usage{
grattan_pal(n = 0, reverse = FALSE)
}
\arguments{
\item{n}{Numeric. The number of levels in your colour scale. Minimum value is 1, maximum is 7.
If you don't specify `n`, a five-colour palette will be used, which may not look right.
Specify `n`. By default, n = 2 will give you light orange and dark orange.
Use n = "2a" if you want light orange and red.}

\item{reverse}{Logical. FALSE by default. Setting to TRUE reverses the
standard colour order. Standard colour order runs from light to dark. If you
set reverse to TRUE, colours will run from dark to light.}
}
\description{
Create a Grattan-appropriate palette for your chart.
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    theme_grattan() +
    scale_colour_manual(values = grattan_pal(n = 3, reverse = FALSE))

p

}
