% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_pal.R
\docType{data}
\name{grattan_palettes}
\alias{grattan_palettes}
\title{Predefined grattan colours combined into palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
grattan_palettes
}
\description{
This is a list of grattan colours combined into palettes. The palettes are used
for different plots and maps.
}
\keyword{datasets}
