% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_point_filled.R
\name{grattan_point_filled}
\alias{grattan_point_filled}
\title{Convenience function to create large points with white centres}
\usage{
grattan_point_filled(size = 3, stroke = 1.5, fill = "white", shape = 21, ...)
}
\arguments{
\item{size}{Size of the point}

\item{stroke}{width of the line around the outside of the point}

\item{fill}{The colour on the inside of the point}

\item{shape}{The shape of the point; default is 21. See
`vignette("ggplot2-specs")` under `Point` for more about this.}

\item{...}{arguments passed to `ggplot2::geom_point()`}
}
\description{
Convenience function to create large points with white centres
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
  grattantheme::grattan_point_filled()
}
