% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_save.R
\name{grattan_save}
\alias{grattan_save}
\title{Save ggplot2 object as an image in the correct size and resolution for Grattan charts. Wrapper around ggsave().}
\usage{
grattan_save(filename, object = ggplot2::last_plot(), type = "normal",
  save_data = FALSE, force_labs = FALSE)
}
\arguments{
\item{filename}{Required. This parameter sets the filename (including path where necessary)
where you want to save your image. The extension defines the file type. Suggested filetypes
are "pdf" or "png", but others are available (see \code{?ggsave} for the full list).}

\item{object}{The ggplot2 graph object to be saved. Defaults to \code{last_plot()}, which
will save the last plot that was displayed in your session.}

\item{type}{Sets height and width to Grattan defaults for one of c("normal", "tiny",
"wholecolumn", "fullpage", "fullslide"). 'normal' is the default and uses default height
and width. 'tiny' uses height of 11.08cm and default width. 'wholecolumn' uses height of
22.16cm and default width. 'fullpage' uses height 22.16cm and width of 44.32cm.
'fullslide' saves an image that can be used as a complete 4:3 Powerpoint slide,
complete with Grattan logo. Set type = "all" to save your chart in all available
sizes.}

\item{save_data}{Logical. Default is FALSE, unless `type` = "all". If set to TRUE, a .csv file will be created containing the dataframe you passed to ggplot(). The filename and path will be the same as your image, but with a .csv extension. Data will always be saved if type = "all".}

\item{force_labs}{Logical. By default, `grattan_save()` will remove your title, subtitle, and caption (if present) from your graph before saving it, unless `type` = \"fullslide\". By setting `force_labs` to TRUE, your title/subtitle/caption will be retained regardless of `type`.}
}
\description{
Save ggplot2 object as an image in the correct size and resolution for Grattan charts. Wrapper around ggsave().
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    theme_grattan()

grattan_save("p.png", p)

# If you don't assign your chart to an object name, that's OK, it will still
# save if it was the last plot displayed.

ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    theme_grattan()

grattan_save("p.png")


# Want to make a 'self-contained' chart that includes a title/subtitle/caption,
# eg. to go on the Grattan Blog? If so, just add them - they'll be properly
# left-aligned when you save them with grattan_save(), like this:

 ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    scale_y_continuous_grattan(limits = c(10, NA)) +
    scale_colour_manual(values = grattan_pal(n = 3)) +
    theme_grattan() +
    labs(title = "Title goes here",
         subtitle = "Subtitle goes here",
         caption = "Notes: Notes go here\\nSource: Source goes here")

 # The plot above won't look right in RStudio's viewer - the text is
 # aligned to the left of the plot area, not the image. Once you save it,
 # the file should have properly-aligned text:

 grattan_save("your_file.png")

 # Want to make a full Powerpoint slide? Just use type = "fullslide" in grattan_save(), like this.
 # If you include 'notes' and 'source' as below, grattan_save() will automatically
 # split them onto separate rows. It will also wrap your title and subtitle
 # automatically over up to 2 rows; and wrap your caption over as many rows
 # as necessary.

 ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    scale_y_continuous_grattan(limits = c(10, NA)) +
    scale_colour_manual(values = grattan_pal(n = 3)) +
    theme_grattan() +
    labs(title = "Title goes here",
         subtitle = "Subtitle goes here",
         caption = "Notes: Notes go here. Source: Source goes here")

 grattan_save("your_file.png", type = "normal")


}
