% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_save.R
\name{grattan_save}
\alias{grattan_save}
\title{Save ggplot2 object as an image in the correct size and resolution for Grattan charts. Wrapper around ggsave().}
\arguments{
\item{filename}{Required. The filename (including path where necessary) for your image on disk. The extension defines the file type.}

\item{object}{Defaults to `last_plot()`. Can specify a different ggplot2 object to be saved.}

\item{height}{Default is 14.5cm, Grattan normal size default. See \code{type}.}

\item{width}{Default is 22.16cm, Grattan normal size default. See \code{type}.}

\item{type}{Sets height and width to Grattan defaults for one of c("normal", "tiny", "wholecolumn", "fullpage"). Normal is the default and uses default height and width. Tiny uses height of 11.08cm and default width. Whole-column uses height of 22.16cm and default width. Full-page uses height 22.16cm and width of 44.32cm.}
}
\description{
Save ggplot2 object as an image in the correct size and resolution for Grattan charts. Wrapper around ggsave().
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    theme_grattan()

grattan_save("p.png", p)
}
