% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_save.R
\name{grattan_save}
\alias{grattan_save}
\title{Save ggplot2 object as an image in the correct size and resolution for Grattan charts. Wrapper around ggsave().}
\arguments{
\item{filename}{Required. The filename (including path where necessary) for your image on disk. The extension defines the file type.}

\item{object}{Defaults to \code{last_plot()}. Can specify a different ggplot2 object to be saved.}

\item{height}{Default is 14.5cm, Grattan normal size default. See \code{type}.}

\item{width}{Default is 22.16cm, Grattan normal size default. See \code{type}.}

\item{type}{Sets height and width to Grattan defaults for one of c("normal", "tiny", "wholecolumn", "fullpage", "fullslide"). 'normal' is the default and uses default height and width. 'tiny' uses height of 11.08cm and default width. 'wholecolumn' uses height of 22.16cm and default width. 'fullpage' uses height 22.16cm and width of 44.32cm. 'fullslide' saves an image that can be used as a complete 4:3 Powerpoint slide, complete with Grattan logo.}
}
\description{
Save ggplot2 object as an image in the correct size and resolution for Grattan charts. Wrapper around ggsave().
}
\examples{
library(ggplot2)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    theme_grattan()

grattan_save("p.png", p)

# If you don't assign your chart to an object name, that's OK, it will still
# save if it was the last plot displayed.

ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    theme_grattan()

grattan_save("p.png")


# Want to make a 'self-contained' chart that includes a title/subtitle/caption,
# eg. to go on the Grattan Blog? If so, just add them - they'll be properly
# left-aligned when you save them with grattan_save(), like this:

 ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    scale_y_continuous_grattan(limits = c(10, NA)) +
    scale_colour_manual(values = grattan_pal(n = 3)) +
    theme_grattan() +
    labs(title = "Title goes here",
         subtitle = "Subtitle goes here",
         caption = "Notes: Notes go here\\nSource: Source goes here")

 # The plot above won't look right in RStudio's viewer - the text is
 # aligned to the left of the plot area, not the image. Once you save it,
 # the file should have properly-aligned text:

 grattan_save("your_file.png")

 # Want to make a full Powerpoint slide? Just use type = "fullslide" in grattan_save(), like:

 ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    scale_y_continuous_grattan(limits = c(10, NA)) +
    scale_colour_manual(values = grattan_pal(n = 3)) +
    theme_grattan() +
    labs(title = "Title goes here",
         subtitle = "Subtitle goes here",
         caption = "Notes: Notes go here\\nSource: Source goes here")

 grattan_save("your_file.png", type = "fullslide")


}
