% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_save_pptx.R
\name{grattan_save_pptx}
\alias{grattan_save_pptx}
\title{Save a ggplot2 chart, or list of charts, as a Powerpoint presentation}
\usage{
grattan_save_pptx(p = ggplot2::last_plot(), filename, type = "fullslide")
}
\arguments{
\item{p}{ggplot2 plot, or a list of ggplot2 plots}

\item{filename}{Filename (including path where necessary) to save your
Powerpoint presentation.}

\item{type}{Chart type. Currently supported types are \code{"fullslide"},
\code{"fullslide_169"}, and \code{"all"}. If you specify multiple types, as
in `type = c("fullslide_169", "fullslide")` or `type = "all"`, multiple
files will be created, with the type added to the filename.}
}
\description{
`grattan_save_pptx()` creates a Grattan-style Powerpoint
presentation with editable vector graphics and editable text. The title,
subtitle, and caption of your graph are placed in the appropriate places.
Speaker notes are included, which include the file path to the R script
from which you called `grattan_save_pptx()`. If you supply a list of plots,
your Powerpoint presentation will include multiple slides -- one per plot.
If you specify multiple types, multiple PPTX files will be created.
}
