% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_scales.R
\name{grattan_scale}
\alias{grattan_colour_manual}
\alias{grattan_fill_manual}
\title{Convenient functions to set Grattan-appropriate palettes}
\usage{
grattan_colour_manual(
  n = 0,
  reverse = FALSE,
  discrete = TRUE,
  faded = FALSE,
  palette = "full",
  ...
)

grattan_fill_manual(
  n = 0,
  reverse = FALSE,
  discrete = TRUE,
  faded = FALSE,
  palette = "full",
  ...
)
}
\arguments{
\item{n}{Numeric. The number of levels in your colour scale. Minimum value is
1, maximum is 7. Passed to \code{grattan_pal}; see ?grattan_pal for more
information.}

\item{reverse}{Logical. FALSE by default. Setting to TRUE reverses the
standard colour order.}

\item{discrete}{Logical. TRUE by default. Setting to FALSE generates a
continuous colour scale.}

\item{faded}{Logical. FALSE by default. Setting to TRUE returns faded
variations of the standard colours.}

\item{palette}{Sets the colours that will form the continuous palette when
  discrete = FALSE. One of:

\itemize{
\item{"full"}{The default. Red, dark orange, light orange, yellow, light
yellow}
\item{"full_f"}{ faded version of "full"}
\item{"light"}{ light
orange, yellow, light yellow}
\item{"dark"}{ red, dark orange, light orange}
\item{"diverging"}{ red, faded red, white, faded light orange, light orange}
\item{"grey"}{ grey 1, grey 2, grey 3, grey 4, grey 5}
}}

\item{...}{arguments passed to ggplot2 scales}
}
\description{
Convenient functions to set Grattan-appropriate palettes
}
\examples{

library(ggplot2)

ggplot(data = mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
   geom_point() +
   grattan_colour_manual(n = 3) +
   theme_grattan()

}
