% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_slide.R
\name{make_slide}
\alias{make_slide}
\title{Create a Powerpoint slide that contains a Grattan graph and editable title and subtitle}
\usage{
make_slide(graph = last_plot(), filename = NULL, path = ".",
  type = "16:9")
}
\arguments{
\item{graph}{The graph you want to include on your slide. Defaults to the last plot (ggplot2::last_plot())}

\item{filename}{The filename for your Powerpoint slide.}

\item{path}{Path to the directory you wish to save your slide in. Defaults to your working directory.}

\item{type}{The type of Powerpoint slide to create; either "16:9" (the default, for a 16:9 orientation slide), or "4:3".}
}
\description{
Create a Powerpoint slide that contains a Grattan graph and editable title and subtitle
}
\examples{

# First, create your ggplot2 object. Include your subtitle, title, and caption with +labs()

library(ggplot2)
library(grattantheme)

ggplot(mtcars, aes(x = mpg, y = hp)) +
   geom_point() +
   theme_grattan() +
   labs(title = "Long title goes here with a bunch of text lorem ipsum
   Grattan blah blah sensible orange",
   subtitle = "Subtitle",
   caption = "Source: blah")

# Now, create a Powerpoint slide with your graph in your working directory.
# By default, make_slide() will use the last plot you generated; you can specify a
# different object to use with the `graph` argument.

make_slide(filename = "test.pptx")

}
