% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_slide.R
\name{make_slide}
\alias{make_slide}
\alias{make_presentation}
\title{Create a Powerpoint slide(s) that contains a Grattan graph and editable title and subtitle.}
\usage{
make_slide(graph = last_plot(), filename = NULL, path = ".", type = "16:9")

make_presentation(
  graphs,
  filename = NULL,
  path = ".",
  type = "16:9",
  title = "Title",
  subtitle = "Subtitle"
)
}
\arguments{
\item{graph}{The graph you want to include on your slide. Defaults to the last plot (ggplot2::last_plot())}

\item{filename}{The filename for your Powerpoint slide.}

\item{path}{Path to the directory you wish to save your slide in. Defaults to your working directory.}

\item{type}{The type of Powerpoint slide to create; either "16:9" (the default, for a 16:9 orientation slide), or "4:3".}

\item{graphs}{A list of ggplot2 objects}

\item{title}{Optional. The title of your presentation, if you want a title slide.}

\item{subtitle}{Optional. The subtitle of your presentation, if you want a title slide.}
}
\description{
Use `make_slide()` to create a Powerpoint presentation containing a single slide,
or `make_presentation()` for a presentation with multiple slides.
}
\examples{

# To take a single ggplot2 graph and turn it into a slide:
# First, create your ggplot2 object. Include your subtitle, title, and caption with +labs()

library(ggplot2)
library(grattantheme)

ggplot(mtcars, aes(x = mpg, y = hp)) +
   geom_point() +
   theme_grattan() +
   labs(title = "Long title goes here with a bunch of text lorem ipsum
   Grattan blah blah sensible orange",
   subtitle = "Subtitle",
   caption = "Source: blah")

# Now, create a Powerpoint slide with your graph in your working directory.
# By default, make_slide() will use the last plot you generated; you can specify a
# different object to use with the `graph` argument.

\donttest{make_slide(filename = "test.pptx")}


# To take multiple ggplot2 graphs and turn them into slides in a single presentation:
# First, create multiple ggplot2 objects. Include your subtitle, title, and caption with +labs()

library(ggplot2)
library(grattantheme)

graph1 <- ggplot(mtcars, aes(x = mpg, y = hp)) +
   geom_point() +
   theme_grattan() +
   labs(title = "Title goes here with a bunch of text lorem ipsum",
   subtitle = "Subtitle",
   caption = "Source: blah")

graph2 <- ggplot(mtcars, aes(x = mpg, y = hp)) +
   geom_col() +
   theme_grattan() +
   labs(title = "This is another graph to go on the second slide",
   subtitle = "Units go here",
   caption = "Source: mtcars.")

# Now combine your two graphs together in a list:

graphs <- list(graph1, graph2)

# Now, create a Powerpoint presentation with one slide per graph in your list:

\donttest{make_presentation(graphs, filename = "test.pptx")}

}
