% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_default_aesthetics.R
\docType{data}
\name{plot_opts_grattan}
\alias{plot_opts_grattan}
\title{A list of plot options reflecting the Grattan brand}
\format{
An object of class \code{list} of length 4.
}
\usage{
plot_opts_grattan
}
\description{
A list of plot options reflecting the Grattan brand
}
\keyword{datasets}
