% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grattan.R
\name{theme_grattan}
\alias{theme_grattan}
\title{Create a ggplot2 theme consistent with the Grattan style guide}
\usage{
theme_grattan(base_size = 18, base_family = "sans")
}
\arguments{
\item{base_size}{Size for text elements. Defaults to 18.}

\item{base_family}{Font family for text elements. Defaults to "sans", indistinguishable from Arial.}
}
\description{
Create a ggplot2 theme consistent with the Grattan style guide
}
\examples{
library(ggplot)
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    theme_grattan()
p
}
