% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grattan.R
\name{theme_grattan}
\alias{theme_grattan}
\title{Create a ggplot2 theme consistent with the Grattan style guide.}
\arguments{
\item{base_size}{Size for text elements. Defaults to 18, as per the Grattan style guide.}

\item{base_family}{Font family for text elements. Defaults to "sans", indistinguishable from Arial.}

\item{flipped}{FALSE by default. Set to TRUE if using coord_flip(). If set to TRUE, the theme will show a vertical axis line, ticks & panel grid, while hiding the horizontals.}

\item{background}{"white" by default. Set to "orange" if you're making a chart to go in a Grattan report box.}
}
\description{
Create a ggplot2 theme consistent with the Grattan style guide.
}
\examples{

# This function goes most of the way to making your charts Grattan-y,
# but manual tweaking will almost definitely be required

# A minimal example:
library(ggplot2)

p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    theme_grattan()
p

# To ensure your x-axis is at zero (or some other value you choose),
# you may need to manually tweak the scale of the y-axis, like this:

p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    scale_y_continuous(limits = c(0, 35),
                       expand = c(0,0)) +
    theme_grattan()
p

# An example with colours follows. See ?grattan_pal for more information.

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    scale_y_continuous(limits = c(0, 35), expand = c(0,0)) +
    scale_colour_manual(values = grattan_pal(n = 3)) +
    theme_grattan()
p

# The legend is off by default. You may wish to turn it on. Here's how:

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
    geom_point() +
    scale_y_continuous(limits = c(0, 35), expand = c(0,0)) +
    scale_colour_manual(values = grattan_pal(n = 3)) +
    theme_grattan() +
    theme(legend.position = "bottom")

p

# The flipped = TRUE option makes things easier when using coord_flip, as in:

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
     geom_point() +
     scale_y_continuous(limits = c(0, 35), expand = c(0,0)) +
     scale_colour_manual(values = grattan_pal(n = 3)) +
     theme_grattan(flipped = TRUE) +
     coord_flip() +
     theme(legend.position = "bottom")

p

# Making a chart to go in a box? Then you'll want the background = "orange" option, as in:

p <- ggplot(mtcars, aes(x = wt, y = mpg, col = factor(cyl))) +
     geom_point() +
     scale_y_continuous(limits = c(0, 35), expand = c(0,0)) +
     scale_colour_manual(values = grattan_pal(n = 3)) +
     theme_grattan(flipped = TRUE, background = "orange") +
     coord_flip() +
     theme(legend.position = "bottom")

p

}
